refactor: node creation and nodes data init
