+++ src/utils/canvas.ts
@@ -1,4 +1,5 @@
-import { Canvas } from "fabric";
+import { Canvas, Rect } from "fabric";
+import { AtlasixDiagram } from "../AtlasixDiagram";
 
 export function createCanvas() {
   let canvas = document.createElement("canvas");
@@ -18,4 +19,33 @@ export function createFabricCanvas(canvas: HTMLCanvasElement) {
   });
 
   return fabricCanvas;
+}
+
+export function createNodesAndSetNodesData(inputData: any[], onSelectedCallback: (e: any, atlasixDiagram: AtlasixDiagram) => void, atlasixDiagram: AtlasixDiagram) {
+  let nodesData: any[] = [];
+
+  for (let [id, node] of inputData.nodes.entries()) {
+    let tmpNode = new Rect({
+      backgroundColor: node.backgroundColor,
+      borderColor: node.borderColor,
+      borderScaleFactor: 2,
+      fill: null,
+      height: node.height,
+      width: node.width,
+      left: node.x,
+      top: node.y
+    });
+    tmpNode.set("id", id);
+    nodesData[id.toString()] = node.data;
+    atlasixDiagram.canvas.add(tmpNode);
+  }
+  
+  atlasixDiagram.canvas.getObjects().forEach((object) => {
+    object.on("selected", (e) => onSelectedCallback(e, atlasixDiagram));
+    object.hasControls = false;
+    object.lockMovementX = true;
+    object.lockMovementY = true;
+  });
+
+  atlasixDiagram.nodesData = nodesData;
 }
